+++ fluent-react/examples/fallback-sync/src/l10n.js
@@ -1,0 +1,21 @@
+import 'fluent-intl-polyfill';
+import { MessageContext } from 'fluent/compat';
+
+const MESSAGES_ALL = {
+  'pl': `
+foo = Foo po polsku
+  `,
+  'en-US': `
+foo = Foo in English
+bar = Bar in English
+  `,
+};
+
+export function* generateMessages() {
+  for (const locale of ['pl', 'en-US']) {
+    const cx = new MessageContext(locale);
+    cx.addMessages(MESSAGES_ALL[locale]);
+    yield cx;
+  }
+}
+