+++ fluent-react/examples/text-input/src/l10n.js
@@ -1,14 +1,8 @@
 import 'fluent-intl-polyfill';
-export { MessagesProvider } from 'fluent-react/compat';
+import { MessageContext } from 'fluent/compat';
 import negotiateLanguages from 'fluent-langneg/compat';
 
-export function negotiateAvailable(requested) {
-  return negotiateLanguages(
-    requested, ['en-US', 'pl'], { defaultLocale: 'en-US' }
-  )
-}
-
-export const MESSAGES_ALL = {
+const MESSAGES_ALL = {
   'pl': `
 hello = Cześć { $username }!
 hello-no-name = Witaj nieznajomy!
@@ -22,3 +16,18 @@ type-name
     .placeholder = Your name
   `,
 };
+
+export function* generateMessages(userLocales) {
+  // Choose locales that are best for the user.
+  const currentLocales = negotiateLanguages(
+    userLocales,
+    ['en-US', 'pl'],
+    { defaultLocale: 'en-US' }
+  );
+
+  for (const locale of currentLocales) {
+    const cx = new MessageContext(locale);
+    cx.addMessages(MESSAGES_ALL[locale]);
+    yield cx;
+  }
+}