+++ fluent-react/test/localized_valid_test.js
@@ -1,0 +1,79 @@
+import React from 'react';
+import assert from 'assert';
+import { shallow, mount } from 'enzyme';
+import { LocalizationProvider, Localized } from '../src/index';
+import Localization from '../src/localization';
+
+suite('Localized - validation', function() {
+  suiteSetup(function() {
+    this.error = console.error;
+    console.error = () => {};
+  });
+
+  suiteTeardown(function() {
+    console.error = this.error;
+  });
+
+  test('inside of a LocalizationProvider', function() {
+    const wrapper = mount(
+      <LocalizationProvider messages={[]}>
+        <Localized>
+          <div />
+        </Localized>
+      </LocalizationProvider>
+    );
+    assert.equal(wrapper.length, 1);
+  });
+
+  test('outside of a LocalizationProvider', function() {
+    function render() {
+      shallow(
+        <Localized />
+      );
+    }
+    assert.throws(render, /descendant of a LocalizationProvider/);
+  });
+
+  test('with a manually set context', function() {
+    const wrapper = shallow(
+      <Localized>
+        <div />
+      </Localized>,
+      { context: { l10n: new Localization([]) } }
+    );
+    assert.equal(wrapper.length, 1);
+  });
+
+  test('without a child', function() {
+    function render() {
+      shallow(
+        <Localized />,
+        { context: { l10n: new Localization([]) } }
+      );
+    }
+    assert.throws(render, /a single React element child/);
+  });
+
+  test('with multiple children', function() {
+    function render() {
+      shallow(
+        <Localized>
+          <div />
+          <div />
+        </Localized>,
+        { context: { l10n: new Localization([]) } }
+      );
+    }
+    assert.throws(render, /a single React element child/);
+  });
+
+  test('without id', function() {
+    const wrapper = shallow(
+      <Localized>
+        <div />
+      </Localized>,
+      { context: { l10n: new Localization([]) } }
+    );
+    assert.equal(wrapper.state('mcx'), null);
+  });
+});