+++ fluent-react/CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## Unreleased
 
-  - …
+  - Removed `MessagesProvider` and replaced it with a new
+    `LocalizationProvider` component.
+
+    The `LocalizationProvider` component takes one prop: `messages`.  It should
+    be an iterable of `MessageContext` instances in order of user's preferred
+    languages.  This iterable will be used by `Localization` to format
+    translations.  If a translation is missing in one language, `Localization`
+    will fall back to the next locale.
+
+  - The relevant `MessageContext` is now cached in `Localized`'s state.
 
 ## fluent-react 0.2.0